# docs/api/tcheck-corporate.openapi.yaml
openapi: 3.0.3
info:
  title: TallyCheck Corporate API
  description: |
    Corporate attendance, SafeChild pickup, BLE beacons and tenant settings.

    This is one of several TallyCheck service contracts. It does **not** cover the
    academic domain (courses, cohorts, enrolment) — that lives in `Tcheck-Backend`
    and is described by `tcheck-academic.openapi.yaml`. Visitor management lives in
    `vcheck.openapi.yaml`. The three are separate documents on purpose: they run on
    different hosts, use different auth, and have colliding paths (`/attendance`
    means corporate clock-in here and class attendance in the academic service).

    ## Multi-tenancy
    Every request resolves to exactly one tenant schema before any handler runs.
    See `shared/tenancy.yaml`. Callers holding a token send nothing extra — the
    token carries the tenant.

    ## Roles
    Roles are enumerated once, in `shared/primitives.yaml#/components/schemas/RoleKey`,
    and must match `libs/auth/src/roles.ts` and `backend/helpers/auth_helper.py`.
    Some tenants still carry the legacy names `hr` and `manager`; the backend
    guards accept both.

    ## Authentication
    All endpoints except `GET /health` and
    `GET /auth/organization-by-subdomain/{subdomain}` require a Bearer JWT.

    ## Response shape
    Successful responses are enveloped as `{ "message": ..., "data": ... }`, with
    `meta` added on paginated collections. See `shared/envelope.yaml`.
  version: 2.1.0
servers:
...
security:
  - Auth0: []

# ==========================================================
# REUSABLE STRUCTURES
# ==========================================================
components:
  securitySchemes:
    Auth0:
      type: openIdConnect
      # Must match the tenant configured in apps/intranet/src/app/app.config.ts.
      # This previously named a different tenant than the app actually used.
      openIdConnectUrl: https://dev-x506paw8afbw6jgu.us.auth0.com/.well-known/openid-configuration

  # ==========================================================
  # REUSABLE SCHEMAS
  # ==========================================================
  schemas:
...
    ErrorResponse:
      $ref: './shared/errors.yaml#/components/schemas/ErrorResponse'
...
    OrgType:
      type: string
      description: >
        Product / vertical. Drives which report shape a tenant renders. `visitor`
        is the vcheck seam.
      enum: [corporate, education, church, visitor]
    OrgStatus:
      type: string
      description: Provisioning lifecycle of a tenant.
      enum: [pending, provisioning, active, failed]
    DashboardKpis:
      type: object
      description: Headline attendance KPIs for a single tenant, today.
      properties:
        total_headcount: { type: integer }
        present_today: { type: integer }
        absent_today: { type: integer }
        currently_clocked_in: { type: integer }
        attendance_rate_today: { type: number, format: float }
        late_arrivals: { type: integer }
        remote_today: { type: integer }
        pending_corrections: { type: integer }
    PlatformOrgRef:
      type: object
      properties:
        id: { type: string, example: org_0D0uiOBBi91ZiW0v }
        name: { type: string }
        domain: { type: string }
        org_type: { $ref: "#/components/schemas/OrgType" }
        status: { $ref: "#/components/schemas/OrgStatus" }
        is_active: { type: boolean }
    PlatformOrgBreakdown:
      allOf:
        - $ref: "#/components/schemas/PlatformOrgRef"
        - type: object
          properties:
            kpis: { $ref: "#/components/schemas/DashboardKpis" }
            error:
              type: string
              nullable: true
              description: Present instead of `kpis` when that tenant could not be read.
    PlatformSummary:
      type: object
      description: Platform-wide totals. Per-org detail is in the envelope `meta`.
      properties:
        total_headcount: { type: integer }
        present_today: { type: integer }
        currently_clocked_in: { type: integer }
        absent_today: { type: integer }
        pending_corrections: { type: integer }
        attendance_rate_today: { type: number, format: float }
        active_organizations: { type: integer }
    PlatformOrgListItem:
      allOf:
        - $ref: "#/components/schemas/PlatformOrgRef"
        - type: object
          properties:
            headcount:
              type: integer
              nullable: true
            present_today:
              type: integer
              nullable: true
    PlatformOrgDashboard:
      type: object
      properties:
        organization: { $ref: "#/components/schemas/PlatformOrgRef" }
        kpis: { $ref: "#/components/schemas/DashboardKpis" }
  responses:
    Unauthorized: # Can be referenced as '#/components/responses/unauthorized'
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /platform/summary:
    get:
      tags: [Platform]
      summary: Platform-wide attendance totals across all active tenants
      description: >
        Super-admin only. Aggregates each active tenant's dashboard KPIs; per-org
        detail is returned in the envelope `meta.organizations`. A tenant that
        cannot be read is reported with an `error` instead of `kpis` rather than
        failing the whole response.
      responses:
        "200":
          description: Aggregate totals + per-org breakdown in meta.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlatformSummary"
                      meta:
                        type: object
                        properties:
                          organizations:
                            type: array
                            items:
                              $ref: "#/components/schemas/PlatformOrgBreakdown"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /platform/organizations:
    get:
      tags: [Platform]
      summary: All organizations with a headline stat each
      description: Super-admin only. Includes suspended orgs; headline stats are null for those.
      responses:
        "200":
          description: Organizations with org_type/status and a cheap headline.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PlatformOrgListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /platform/organizations/{org_id}/dashboard:
    get:
      tags: [Platform]
      summary: One organization's dashboard, viewed by a super-admin
      description: >
        Super-admin only. Runs the identical dashboard query the tenant's own
        admins see, against that tenant's schema.
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
          example: org_0D0uiOBBi91ZiW0v
      responses:
        "200":
          description: The organization plus its dashboard KPIs.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PlatformOrgDashboard"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"