# libs/api-client/src/generated/tcheck/model/platformOrgDashboard.ts
/**
 * Generated by orval from TallyCheck Corporate API v2.1.0.
 * DO NOT EDIT — run `nx run api-client:generate` instead.
 */
import type { DashboardKpis } from "./dashboardKpis";
import type { PlatformOrgRef } from "./platformOrgRef";

export interface PlatformOrgDashboard {
  organization?: PlatformOrgRef;
  kpis?: DashboardKpis;
}