# libs/api-client/src/generated/tcheck/model/platformOrgListItem.ts
/**
 * Generated by orval from TallyCheck Corporate API v2.1.0.
 * DO NOT EDIT — run `nx run api-client:generate` instead.
 */
import type { PlatformOrgRef } from "./platformOrgRef";

export type PlatformOrgListItem = PlatformOrgRef & {
  /** @nullable */
  headcount?: number | null;
  /** @nullable */
  present_today?: number | null;
};